The task am_i_wilson was updated

+++ src/main/java/org/academy/kata/implementation/Vovanchek/Eight.java
@@ -45,7 +45,7 @@ public class Eight implements IEight {
     }
 
     @Override
-    public boolean am_i_wilson(double n) {
+    public boolean am_i_wilson(long n) {
          if (n <= 1) {
             return false;
         }
